Add battery level logging.
